fix: coverage_months поля не сериализовались из-за Decimal — заменён cast на Float

+++ backend/src/services/sale/primary.py
@@ -496,7 +496,7 @@ class PrimarySalesAndStockService(
                     ),
                     else_=None,
                 ),
-                Numeric(10, 2),
+                Float,
             ).label("coverage_months_amount"),
             func.cast(
                 case(
@@ -507,7 +507,7 @@ class PrimarySalesAndStockService(
                     ),
                     else_=None,
                 ),
-                Numeric(10, 2),
+                Float,
             ).label("coverage_months_packages"),
         ]
